Backend/requirements.txt
#run using : pip install -r requirements.txt
# add all the imports that need to be installed here

# flask==3.1.0           
# flask-jwt-extended==4.7.1                    
# flask-login==0.6.3           
# flask-sqlalchemy==3.1.1           
# python-dotenv==1.1.0               
# requests==2.32.3
# sqlalchemy==2.0.37
# sqlite==3.45.3  
# argon2-cffi==23.1.0
# flask-socketio
# eventlet
# flask-cors
# passlib



# # Core Flask and Web Server
# Flask==3.1.0
# gunicorn==21.2.0
# Werkzeug==3.1.3
# argon2-cffi==23.1.0

# # Database
# Flask-SQLAlchemy==3.1.1
# psycopg2-binary==2.9.10

# # Authentication and Security
# Flask-Login==0.6.3
# Flask-JWT-Extended==4.7.1
# passlib==1.7.4
# PyJWT==2.10.1
# cryptography==44.0.2

# # Firebase
# firebase-admin==6.7.0
# google-api-python-client==2.166.0
# google-auth==2.38.0
# google-cloud-firestore==2.20.1
# google-cloud-storage==3.1.0

# # WebSockets
# Flask-SocketIO==5.5.1
# python-socketio==5.13.0
# python-engineio==4.12.0

# # Utilities
# python-dotenv==1.1.0
# requests==2.32.3
# Flask-Cors==5.0.1
# Flask-Limiter==3.12
# limits==5.0.0


# Core Flask and Web Server
Flask==3.1.0
Werkzeug==3.1.3
gunicorn==21.2.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-limiter==3.12
flask-login==0.6.3
flask-socketio==5.5.1
flask-sqlalchemy==3.1.1

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.37

# Authentication and Security
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
passlib==1.7.4
PyJWT==2.10.1
cryptography==44.0.2
cachetools==5.5.2

# Firebase Services
firebase-admin==6.7.0
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-api-core==2.24.2
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
proto-plus==1.26.1
protobuf==5.29.4

# WebSockets and Async
python-socketio==5.13.0
python-engineio==4.12.0
simple-websocket==1.1.0

# Utilities
python-dotenv==1.1.0
requests==2.32.3
httplib2==0.22.0
urllib3==2.3.0
pysocks==1.7.1
uritemplate==4.1.1
rich==13.9.4
pygments==2.19.1

# FastAPI (if used)
fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2
h11==0.16.0
anyio==4.9.0

# Development (optional)
pydantic==2.11.4
pydantic-core==2.33.2
typing-extensions==4.12.2
typing-inspection==0.4.0
annotated-types==0.7.0